analyzer: Move expandShortcutURL() to PackageJsonUtils
Do accordingly for the tests to get rid of the "NpmTest" unit test as we
have a functional test of the same name, which was confusing.
