--- analyzer/src/test/kotlin/NpmTest.kt
@@ -1,69 +1,0 @@
-/*
- * Copyright (C) 2017-2019 HERE Europe B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-package com.here.ort.analyzer
-
-import com.here.ort.analyzer.managers.Npm
-
-import io.kotlintest.shouldBe
-import io.kotlintest.specs.WordSpec
-
-class NpmTest : WordSpec({
-    "expandShortcutURL" should {
-        "do nothing for empty URLs" {
-            Npm.expandShortcutURL("") shouldBe ""
-        }
-
-        "properly handle NPM shortcut URLs" {
-            val packages = mapOf(
-                "npm/npm"
-                        to "https://github.com/npm/npm.git",
-                "mochajs/mocha#4727d357ea"
-                        to "https://github.com/mochajs/mocha.git#4727d357ea",
-                "github:snyk/node-tap#540c9e00f52809cb7fbfd80463578bf9d08aad50"
-                        to "https://github.com/snyk/node-tap.git#540c9e00f52809cb7fbfd80463578bf9d08aad50",
-                "gist:11081aaa281"
-                        to "https://gist.github.com/11081aaa281",
-                "bitbucket:example/repo"
-                        to "https://bitbucket.org/example/repo.git",
-                "gitlab:another/repo"
-                        to "https://gitlab.com/another/repo.git"
-            )
-
-            packages.forEach { (actualUrl, expectedUrl) ->
-                Npm.expandShortcutURL(actualUrl) shouldBe expectedUrl
-            }
-        }
-
-        "not mess with crazy URLs" {
-            val packages = mapOf(
-                "git@github.com/cisco/node-jose.git"
-                        to "git@github.com/cisco/node-jose.git",
-                "https://git@github.com:hacksparrow/node-easyimage.git"
-                        to "https://git@github.com:hacksparrow/node-easyimage.git",
-                "github.com/improbable-eng/grpc-web"
-                        to "github.com/improbable-eng/grpc-web"
-            )
-
-            packages.forEach { (actualUrl, expectedUrl) ->
-                Npm.expandShortcutURL(actualUrl) shouldBe expectedUrl
-            }
-        }
-    }
-})
