+++ analyzer/src/test/kotlin/PackageJsonUtilsTest.kt
@@ -19,6 +19,7 @@
 
 package com.here.ort.analyzer
 
+import com.here.ort.analyzer.PackageJsonUtils.Companion.expandShortcutURL
 import com.here.ort.analyzer.PackageJsonUtils.Companion.mapDefinitionFilesForNpm
 import com.here.ort.analyzer.PackageJsonUtils.Companion.mapDefinitionFilesForYarn
 import com.here.ort.utils.safeDeleteRecursively
@@ -145,6 +146,48 @@ class PackageJsonUtilsTest : WordSpec() {
 
                 mapDefinitionFiles(definitionFiles) shouldContainExactlyInAnyOrder
                         absolutePaths("a/package.json", "a/b/c/e/package.json")
+            }
+        }
+
+        "expandShortcutURL" should {
+            "do nothing for empty URLs" {
+                expandShortcutURL("") shouldBe ""
+            }
+
+            "properly handle NPM shortcut URLs" {
+                val packages = mapOf(
+                    "npm/npm"
+                            to "https://github.com/npm/npm.git",
+                    "mochajs/mocha#4727d357ea"
+                            to "https://github.com/mochajs/mocha.git#4727d357ea",
+                    "github:snyk/node-tap#540c9e00f52809cb7fbfd80463578bf9d08aad50"
+                            to "https://github.com/snyk/node-tap.git#540c9e00f52809cb7fbfd80463578bf9d08aad50",
+                    "gist:11081aaa281"
+                            to "https://gist.github.com/11081aaa281",
+                    "bitbucket:example/repo"
+                            to "https://bitbucket.org/example/repo.git",
+                    "gitlab:another/repo"
+                            to "https://gitlab.com/another/repo.git"
+                )
+
+                packages.forEach { (actualUrl, expectedUrl) ->
+                    expandShortcutURL(actualUrl) shouldBe expectedUrl
+                }
+            }
+
+            "not mess with crazy URLs" {
+                val packages = mapOf(
+                    "git@github.com/cisco/node-jose.git"
+                            to "git@github.com/cisco/node-jose.git",
+                    "https://git@github.com:hacksparrow/node-easyimage.git"
+                            to "https://git@github.com:hacksparrow/node-easyimage.git",
+                    "github.com/improbable-eng/grpc-web"
+                            to "github.com/improbable-eng/grpc-web"
+                )
+
+                packages.forEach { (actualUrl, expectedUrl) ->
+                    expandShortcutURL(actualUrl) shouldBe expectedUrl
+                }
             }
         }
     }
